neurotrace_backend/requirements.txt
flask
flask-jwt-extended
flask-sqlalchemy
flask-migrate
flask-cors
flask-bcrypt

# ML & Data
numpy
pandas
scikit-learn
joblib
shap
lime

# Audio / Voice feature extraction
librosa
soundfile
praat-parselmouth

# Image processing (MRI + Spiral)
Pillow
opencv-python-headless
tensorflow     # only if MRI/Spiral models are TF-based

# Utils
python-dotenv
marshmallow
pydantic
werkzeug
gunicorn
redis         # for rate limiting / caching